python3: use the io module everywhere
Make the calibre plugin usage work on python3. Since FanFicFare does not
support python 2.5, make it work by using the modern idiom.

Essentially, six.StringIO and six.BytesIO makes no sense to use
anywhere, since any code that works on python3 at all will also work
with the io module in python >= 2.6. The only caveat is that the
unadorned str type in python2 works with neither, but it is always best
to be explicit and use either unicode or bytes.