optimization: do not use filter just to get counts of a list
This breaks in python3, because filter returns an iterable filter
object, and it's anyways wrong on python2, because it does extra work to
create a list only to discard it. The best way to do this is to test
each member of the list and abort early as soon as we get the info we
need (that it is non-zero). Or alternatively use sum with a generator
expression.