Pass None to bs4.find() instead of empty string
BeatifulSoup apparently treats an empty string like any other string,
i.e. as a search filter for tag names. Passing None seems to disable
this filtering. Couldn't find the relevant docs for this, so any
representation of find()'s behaviour here is based on my testing only.
(on BeatifulSoup 4.9.3)

A quick code search on the repo shows this is the only place find gets
explicitly passed an empty string. Indeed, all other usages seem to list
the tag to expect (in this case, both are <div>s), but I chose to follow
this adapter's choice of not filtering.