Detect when the background job SIGABRT and don't iter on None
This a workaround for an issue in Calibre C++ code[1] when use_browser_cache and
open_pages_in_browser are set to true.
At the end of the job, calibre-parallel abort in QScopedPointerDeleter<QPlatformServices>::cleanup.
We would see in the log:
FFF: INFO: 2025-06-22 14:30:42,438: writer_epub.py(395): Saving EPUB Version 2.0
free(): invalid next size (fast)

We cannot do anything about that in FanFicFare but we can detect when that happen because
job.result will be None.
And we will trigger an TypeError when generating the job summary:

Traceback (most recent call last):
  File "/usr/bin/calibre-parallel", line 21, in <module>
    sys.exit(main())
             ~~~~^^
  File "/usr/lib64/calibre/calibre/utils/ipc/worker.py", line 214, in main
    result = func(*args, **kwargs)
  File "/usr/lib64/calibre/calibre/utils/ipc/worker.py", line 149, in arbitrary_n
    return func(*args, **kwargs)
  File "calibre_plugins.fanficfare_plugin.jobs", line 139, in do_download_worker_multiproc
    book_list.extend(job.result)
    ~~~~~~~~~~~~~~~~^^^^^^^^^^^^
TypeError: 'NoneType' object is not iterable

This commit fake a failed job result so the job summary will show:
- Status: Error
- Comment: Synthetic failure generated by FFF, check Calibre log

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2374170

Signed-off-by: Nicolas SAPA <nico@ByMe.at>